pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "cdp-mcp"
version = "0.1.0"
description = "MCP server for Cloudera/CDP cluster administration and troubleshooting. Fork of dvergari/cloudera-mcp-server."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"

authors = [
    { name = "Daniele Isoardi" },
    { name = "Davide Vergari" },
]

keywords = ["cloudera", "cdp", "mcp", "llm", "ai", "devops"]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Monitoring",
]

dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "tenacity>=8.3.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "structlog>=24.2.0",
    "python-dotenv>=1.0.0",
    "python-dateutil>=2.9.0",
    "pyyaml>=6.0.0",
    # Iceberg/Impala registry backend (optional at runtime)
    "impyla>=0.21.0",
    "thrift>=0.16.0",
    "thrift-sasl>=0.4.3",
    "pure-sasl>=0.6.2",
]

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0.0"
pytest-asyncio = ">=0.23.0"
pytest-cov = ">=5.0.0"
respx = ">=0.21.0"
ruff = ">=0.4.0"
mypy = ">=1.10.0"

[project.scripts]
cdp-mcp = "cdp_mcp.server:run"

[tool.poetry]
packages = [{ include = "cdp_mcp", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.11"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
